Add ConfirmBookingCompomemt file and add Navbar for BookingSuccesful Page, add ConfirmbookingPage in pages folder and change path of ConfirmBookingSuccess in main.jsx
