+++ src/pages/ConfirmBookingPage.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import Navbar from "../components/Navbar";
+import ConfirmBookingPopup from "../components/ConfirmBookingComponent/confirmBookingPopup";
+
+function ConfirmBookingPage() {
+    return (
+        <div>
+            <Navbar/>
+            <ConfirmBookingPopup/>
+        </div>
+    )
+}
+
+export default ConfirmBookingPage;
